refactor(zql): extract query execution into separate run.ts module
- Extract runQuery, materializeQuery, and preloadQuery functions to new run.ts
- Remove delegateSymbol mechanism from Query interface and QueryImpl
- Update Zero client to use bare execution functions instead of method chaining
- Add completedAST() method to Query interface for accessing completed AST
- Preserve existing query.run(), query.materialize(), and query.preload() methods
  by delegating to new bare functions for backward compatibility
- Update integration test to use runQuery function directly

This refactoring separates query building from query execution, creating
cleaner separation of concerns while maintaining all existing functionality.
