+++ packages/zql-integration-tests/src/v20-array-style.test.ts
@@ -10,7 +10,7 @@ import {
   table,
 } from '../../zero-schema/src/builder/table-builder.ts';
 import {createBuilder} from '../../zql/src/query/named.ts';
-import {delegateSymbol} from '../../zql/src/query/query.ts';
+import {runQuery} from '../../zql/src/query/run.ts';
 
 test('reading from old array style tables', async () => {
   const lc = createSilentLogContext();
@@ -46,7 +46,7 @@ test('reading from old array style tables', async () => {
   const d = newQueryDelegate(lc, testLogConfig, sqlite, schema);
   const queries = createBuilder(schema);
 
-  const rows = await queries.bar[delegateSymbol](d).run();
+  const rows = await runQuery(d, queries.bar);
   expect(rows).toMatchInlineSnapshot(`
     [
       {
